Fix storing pointer difference as 32-bit integer in clipmap loader

+++ src/qcommon/cm_local.h
@@ -98,10 +98,10 @@ typedef struct {
 	int cluster;
 	int area;
 
-	int firstLeafBrush;
+	ptrdiff_t firstLeafBrush;
 	int numLeafBrushes;
 
-	int firstLeafSurface;
+	ptrdiff_t firstLeafSurface;
 	int numLeafSurfaces;
 } cLeaf_t;
